Item Drops Proper Commit
TortoiseGit apparently did not commit properly so I am using Github Desktop quickly to fix this

+++ Mouthwatering Monstrosities/Assets/Scripts/ItemDrops.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 public class itemDrops : ScriptableObject
 {
     public GameObject dropModel;
-    public float dropChance;
+    public int dropChance;
     public string itemName;
 
-    public itemDrops(string lootName, float dropChance) {
+    public itemDrops(string lootName, int dropChance) {
         this.itemName = lootName;
         this.dropChance = dropChance;
     }
